Bugfix with cyclical types & return types
Cyclical types used to not halt, but that's no longer an issue. It has been resolved by assuming that a type under the same alias is always equal.

early return detection works property rather than just changing every if statement with no else block to contain the rest of the body in the current scope.