Fixed assignment to finalized types & types from type defs
Types from type definitions are specified to deduced function return types when no return type is specified.

Almost fixed a bug with assigning to finalized types that are exist in terms of unions. Now, instead of just using sameTypesNoUnionSpec or sameTypes, sameTypesImpl is used which depending on the given rhs either returns the former or the latter. This will later be extended to check for deep unions.