Bugfix with circular dependencies in inference
if x's type depended on y and y's type depended on x where x is declared before y and mutated with a different type after y, then y's type remained unchanged but now that's no longer the case. Unless it relates to functions since they capture the local scope as it is.