Semantic support for arrays + generalize length of newtype
It was unintentionally impossible to generalize over the length of types, so now you can send any number of arguments to a newtype as long as it leaves no variables original generics (This check is yet to be implemented) uninstantiated.

Also added support for arrays at a semantic level, they are fully usable but also very inconvenient because there's no syntactic support for them. The best part is that they are fully expressed in this language's type system. An example of this is available in unionTest.junu along with an example of linked lists.