New machinery for refinements of unions
This comes with new semantics for an `is` operation, such that it now actively filters the set of types too keep by matching every type against them. The new machinery is a little expression language and because of that, it's much more expressive and now allows for expressions such as not to appear in union refinements.