Less directionally biased addTypeVariable unification
This entailed fixing specification from unions to unions to be using sameTypes although it seems like it should always have been using that since the beginning anyway.

Now it sorts old value and new value of a type variable when adding to it, by the amount of generics(excluding rigids) each if them include.